+++ server/src-lib/Hasura/SQL/Types.hs
@@ -4,6 +4,7 @@ module Hasura.SQL.Types
   ( ToSQL (..),
     toSQLTxt,
     CollectableType (..),
+    ExtensionsSchema (..),
   )
 where
 
@@ -52,3 +53,7 @@ instance (ToSQL a) => ToSQL (CollectableType a) where
     CollectableTypeScalar ty -> toSQL ty
     -- typename array is an sql standard way of declaring types
     CollectableTypeArray ty -> toSQL ty <> " array"
+
+-- | The name of the schema in which the graphql-engine will install database extensions.
+newtype ExtensionsSchema = ExtensionsSchema {_unExtensionsSchema :: Text}
+  deriving (Show, Eq, FromJSON, ToJSON, Hashable, Cacheable, NFData)